+++ ICD.Profound.TouchCUE/Themes/TouchCueTheme.cs
@@ -1,8 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using ICD.Common.Utils;
 using ICD.Common.Utils.Collections;
-using ICD.Common.Utils.Extensions;
 using ICD.Connect.Themes.UserInterfaceFactories;
 using ICD.Profound.ConnectPROCommon.Themes;
 using ICD.Profound.TouchCUE.Themes.TouchDisplayInterface;
@@ -12,7 +10,6 @@ namespace ICD.Profound.TouchCUE.Themes
 	public sealed class TouchCueTheme : AbstractConnectProTheme<TouchCueThemeSettings>
 	{
 		private readonly IcdHashSet<IUserInterfaceFactory> m_UiFactories;
-		private readonly SafeCriticalSection m_UiFactoriesSection;
 
 		/// <summary>
 		/// Constructor.
@@ -23,42 +20,14 @@ namespace ICD.Profound.TouchCUE.Themes
 			{
 				new ConnectProTouchDisplayInterfaceFactory(this)
 			};
-
-			m_UiFactoriesSection = new SafeCriticalSection();
 		}
-		
-
-		#region Public Methods
 
 		/// <summary>
 		/// Gets the UI Factories.
 		/// </summary>
 		public override IEnumerable<IUserInterfaceFactory> GetUiFactories()
 		{
-			return m_UiFactoriesSection.Execute(() => m_UiFactories.ToArray())
-			                           .Concat(base.GetUiFactories());
+			return m_UiFactories.Concat(base.GetUiFactories());
 		}
-
-		/// <summary>
-		/// Clears the instantiated user interfaces.
-		/// </summary>
-		public override void ClearUserInterfaces()
-		{
-			base.ClearUserInterfaces();
-
-			m_UiFactoriesSection.Execute(() => m_UiFactories.ForEach(f => f.Clear()));
-		}
-
-		/// <summary>
-		/// Clears and rebuilds the user interfaces.
-		/// </summary>
-		public override void BuildUserInterfaces()
-		{
-			base.BuildUserInterfaces();
-
-			m_UiFactoriesSection.Execute(() => m_UiFactories.ForEach(f => f.BuildUserInterfaces()));
-		}
-
-		#endregion
 	}
 }