+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Fixed a bug where certain UIs were not being instantiated
+
 ## [13.0.0] - 2020-04-29
 ### Added
  - Added TouchCUE UI